Add max frame delta
Max frame delta for the time-dependent portions of the engine is now 1 second. So if the framerate of a game drops below 1 FPS the engine will actually slow down. I don't think this is a big limitation though... if the framerate is that low you have bigger problems.

Did this because pausing execution while debugging was producing frame deltas tens of seconds long and that was breaking animations and other things that try to compensate for low framerate situations.