Enhance scene/dialog management
Moved scene management functions to their own file and overhauled how scenes and dialogs are handled. Scenes no longer open and contain dialogs as "subscenes", instead the engine handles dialogs itself.

Opening a dialog now disables input events for the scene that opened it, which solves a bunch of little issues and I'm embarrassed that it took me so long to think to do this. A dialog that opens another dialog also gets its inputs suppressed; only the most recently opened dialog receives input.

All scenes/dialogs continue to be updated while other dialogs are open. In a later commit I'll add the ability to pause a scene and keep it from updating, and we'll add that logic to the scene management we've set up here.

Also made the main console/compositor into a proper window element, which open up some nice functionality for later and allows us to handle multiple dialogs simultaneously.

Not a bad commit for being away for 2 months! Hopefully it holds up.