Add List scrolling, rework List item selection
Added functions for scrolling the contents of a list. Just some simple ones for now, more fine-grained control for scrolling can be added later. I also haven't added scroll actions yet... not sure how I want that to work just yet. User can obviously do it themselves.

Also implemented the idea of a list being just a list, or a list that allows for an item to be selected. When selection = true, the list will have a selected item and the item will always be kept visible. This was the default behaviour before but not all lists need item selection abilities.

Still to figure out: delineating between scrolling and changing selection, configuring input actions for the 2 modes. Hmmm think think think.